+++ TelegramBot/src/main/java/org/menus/MenuState.java
@@ -1,11 +1,11 @@
-package org.users.states;
+package org.menus;
 
 import org.users.TelegramUser;
 
 public abstract class MenuState {
     abstract public String respondTo(String message);
     abstract public String getQuestion();
-    TelegramUser user;
+    protected TelegramUser user;
     public MenuState(TelegramUser user){
         this.user = user;
     }
