+++ TelegramBot/src/main/java/org/menus/userMenu/RegisterUserMenu.java
@@ -1,6 +1,7 @@
-package org.users.states;
+package org.menus.userMenu;
 
 import org.users.TelegramUser;
+import org.menus.MenuState;
 
 public class RegisterUserMenu extends MenuState {
     @Override
